Applied temp fix to typecasting bug in testsuite.
Details:
- Applied a temporary fix to the typecasting bug in the testsuite driver.
  The fix involves casting both numerator and denominator to unsigned long.
  This fix is more voodoo than science, as I can't be sure why it even
  works.

+++ testsuite/src/test_libblis.c
@@ -729,8 +729,9 @@ dim_t libblis_test_get_dim_from_prob_size( int   dim_spec,
 {
 	dim_t dim;
 
-	if ( dim_spec < 0 ) dim = p_size / bli_abs(dim_spec);
-	else                dim = dim_spec;
+	if ( dim_spec < 0 ) dim = ( unsigned long )p_size /
+	                          ( unsigned long )bli_abs(dim_spec);
+	else                dim = ( unsigned long )dim_spec;
 
 	return dim;
 }
